Put bmp_finder in separate file
